added check for unzipped bundle size

+++ src/commands/deploy.ts
@@ -7,7 +7,7 @@ import { BundlerComposer } from "../bundlers/bundlerComposer.js";
 import { DartBundler } from "../bundlers/dart/dartBundler.js";
 import { NodeJsBinaryDependenciesBundler } from "../bundlers/node/nodeJsBinaryDependenciesBundler.js";
 import { NodeJsBundler } from "../bundlers/node/nodeJsBundler.js";
-import { REACT_APP_BASE_URL } from "../constants.js";
+import { BUNDLE_SIZE_LIMIT, REACT_APP_BASE_URL } from "../constants.js";
 import {
   GENEZIO_NOT_AUTH_ERROR_MSG,
   GENEZIO_NO_CLASSES_FOUND
@@ -26,6 +26,7 @@ import {
   directoryContainsIndexHtmlFiles,
   directoryContainsHtmlFiles,
   deleteFolder,
+  checkBundleFolderSizeLimit,
 } from "../utils/file.js";
 import { printAdaptiveLog, debugLogger } from "../utils/logging.js";
 import { runNewProcess } from "../utils/process.js";
@@ -282,8 +283,17 @@ export async function deployClasses(configuration: YamlProjectConfiguration, clo
         `genezioDeploy.zip`
       );
 
+      // check if the unzipped folder is smaller than 250MB
+      const bundlerSizeCheck = await checkBundleFolderSizeLimit(output.path, BUNDLE_SIZE_LIMIT);
+      debugLogger.debug(`The bundlerSizeCheck for class ${element.path} is ${bundlerSizeCheck}.`);
+      if (!bundlerSizeCheck) {
+        throw new Error(`The size of the bundled folder for class ${element.path} is bigger than ${BUNDLE_SIZE_LIMIT/1048576}MB.`);
+      }
+
       debugLogger.debug(`Zip the directory ${output.path}.`);
       await zipDirectory(output.path, archivePath);
+
+
 
       await deleteFolder(output.path);
